Ts exports attempt
