+++ .gitignore
@@ -14,3 +14,4 @@ docs/
 .env
 build
 node_modules/
+ffi_helpers/lib
