+++ ffi_helpers/poseidonHash.js
@@ -1,22 +1,7 @@
-const { ethers } = require("ethers");
-const { BigNumber } = require("ethers");
-const { buildPoseidon } = require("circomlibjs");
+const { hash } = require('./lib/poseidonHash')
 
-function poseidonHash(poseidon, inputs) {
-  const hash = poseidon(inputs.map((x) => BigNumber.from(x).toBigInt()));
-  // Make the number within the field size
-  const hashStr = poseidon.F.toString(hash);
-  // Make it a valid hex string
-  const hashHex = BigNumber.from(hashStr).toHexString();
-  // pad zero to make it 32 bytes, so that the output can be taken as a bytes32 contract argument
-  const bytes32 = ethers.utils.hexZeroPad(hashHex, 32);
-  return bytes32;
-}
-
-async function hash(left, right) {
-  let poseidon = await buildPoseidon();
-  let res = poseidonHash(poseidon, [left, right]);
-  console.log(res);
-}
-
-hash(process.argv[2], process.argv[3]);
+hash(process.argv[2], process.argv[3])
+  .then((res) => {
+    console.log(res)
+    process.exit(0)
+  });
