+++ ffi_helpers/tree.js
@@ -1,59 +1,7 @@
-const { MerkleTree } = require("./merkleTree.js");
-const { ethers } = require("ethers");
+const { tree } = require('./lib/tree')
 
-const { BigNumber } = require("ethers");
-const { buildPoseidon } = require("circomlibjs");
-
-function poseidonHash(poseidon, inputs) {
-  const hash = poseidon(inputs.map((x) => BigNumber.from(x).toBigInt()));
-  // Make the number within the field size
-  const hashStr = poseidon.F.toString(hash);
-  // Make it a valid hex string
-  const hashHex = BigNumber.from(hashStr).toHexString();
-  // pad zero to make it 32 bytes, so that the output can be taken as a bytes32 contract argument
-  const bytes32 = ethers.utils.hexZeroPad(hashHex, 32);
-  return bytes32;
-}
-
-class PoseidonHasher {
-  poseidon;
-
-  constructor(poseidon) {
-    this.poseidon = poseidon;
-  }
-
-  hash(left, right) {
-    return poseidonHash(this.poseidon, [left, right]);
-  }
-}
-
-async function tree(height, _pushedCommitments, newCommitment) {
-  let poseidon = await buildPoseidon();
-
-  const tree = new MerkleTree(height, "test", new PoseidonHasher(poseidon));
-
-  const pushedCommitments = ethers.utils.defaultAbiCoder.decode(
-    ["bytes32[]"],
-    _pushedCommitments,
-  )[0];
-
-  for (let i = 0; i < pushedCommitments.length; i++) {
-    await tree.insert(pushedCommitments[i]);
-  }
-
-  const before = await tree.root();
-
-  await tree.insert(newCommitment);
-
-  const elements = tree.totalElements;
-  const after = await tree.root();
-
-  console.log(
-    ethers.utils.defaultAbiCoder.encode(
-      ["bytes32", "uint256", "bytes32"],
-      [before, elements, after],
-    ),
-  );
-}
-
-tree(parseInt(process.argv[2]), process.argv[3], process.argv[4]);
+tree(parseInt(process.argv[2]), process.argv[3], process.argv[4])
+  .then((res) => {
+    console.log(res)
+    process.exit(0)
+  });
