+++ helpers/getPoseidonBytecode.ts
@@ -1,0 +1,3 @@
+import { poseidonContract } from 'circomlibjs'
+
+export const poseidon = (nInputs: number) => poseidonContract.createCode(nInputs)
