+++ package.json
@@ -8,7 +8,8 @@
     "test": "test"
   },
   "scripts": {
-    "build": "npm run compile:circuits && npm run setup && npm run compile",
+    "build": "npm run compile:circuits && npm run setup && npm run js-build && npm run compile",
+    "js-build": "tsc",
     "compile": "forge build",
     "compile:circuits": "mkdir -p build && circom circuits/withdraw.circom --r1cs --wasm -o build && circom circuits/deposit.circom --r1cs --wasm -o build",
     "setup": "sh quickSetup.sh",
@@ -24,5 +25,11 @@
     "circomlibjs": "^0.1.7",
     "ethers": "^5.5.4",
     "snarkjs": "^0.7.1"
+  },
+  "devDependencies": {
+    "@types/circomlibjs": "^0.1.3",
+    "@types/node": "^20.8.2",
+    "@types/snarkjs": "^0.7.4",
+    "typescript": "^5.2.2"
   }
 }
